Move TJestMock from generic types to unit test types; Add mock for api-config module;
