+++ __mocks__/bcrypt.ts
@@ -1,4 +1,4 @@
-import type { TJestMock } from '../src/types';
+import type { TJestMock } from '../test/unit/test-index';
 
 const bcrypt: TJestMock & { compare: TJestMock; hash: TJestMock } = jest.createMockFromModule('bcrypt');
