+++ __mocks__/jsonwebtoken.ts
@@ -1,4 +1,4 @@
-import { TJestMock } from '../src/types';
+import { TJestMock } from '../test/unit/test-index';
 
 const jwt: TJestMock & {
   verify: TJestMock;
