+++ __mocks__/mongodb.ts
@@ -1,4 +1,4 @@
-import { TJestMock } from '../src/types';
+import { TJestMock } from '../test/unit/test-index';
 
 const mongodb: TJestMock & { ObjectId: () => string } = jest.createMockFromModule('mongodb');
 mongodb.ObjectId = function () {
