+++ src/database/__mocks__/database-index.ts
@@ -1,4 +1,4 @@
-import { TJestMock } from '../../types';
+import { TJestMock } from '../../../test/unit/test-index';
 import { getMockImplementationError } from '../../../test/unit/mockUtils';
 
 type TMockWithProps = TJestMock & Partial<{ _succeededCall: any; _failedCall: any }>;
