+++ src/middleware/routes/__mocks__/api-config.ts
@@ -1,0 +1,3 @@
+const apiConfigMock = Object.freeze({ api: 'config' });
+
+export default apiConfigMock;
