+++ test/unit/test-index.ts
@@ -32,3 +32,5 @@ const mockAndRequireModule = (path: string) => {
 };
 
 export { findAssociatedSrcModulePath, getRootRelativePath, mockAndRequireModule };
+
+export type TJestMock<T = void> = jest.Mock<any, T | any>;
